+++ src/component/App.tsx
@@ -1,0 +1,15 @@
+import './App.css'
+import Header  from './header/header'
+
+function App() {
+  return (
+    <div className="main">
+      <Header />
+      <div className="container">
+       
+      </div>
+    </div>
+  );
+}
+
+export default App;
