+++ src/component/header/header.css
@@ -1,0 +1,49 @@
+.header {
+  display: flex;
+  justify-content: center;
+  padding: 16px 0;
+}
+
+.nav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  max-width: 1200px;
+}
+
+.nav-links {
+  display: flex;
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  gap: 8px;
+}
+
+.nav-link {
+  display: flex;
+  align-items: center;
+  padding: 16px;
+  gap: 9px;
+}
+
+.nav-link p {
+  color: white;
+}
+
+.logo-container {
+  display: flex;
+  flex-grow: 1;
+  justify-content: center;
+}
+
+.account {
+  display: flex;
+  align-items: center;
+  padding: 16px;
+  gap: 9px;
+}
+
+.account p {
+  color: white;
+}
